tagfilter_service/requirements.txt
Flask>=2.2.0,<3.0.0
flask-cors
nltk>=3.7
openai>=0.27.0
pymongo>=4.0.0,<5.0.0
python-dotenv>=0.21.0
spacy==3.7.2
thinc==8.2.2
numpy==1.24.4
dnspython
setuptools>=65.0.0
